# Conflicts:
#	src/main/java/ie/atu/userinterface/Admin/AdminControllerNew.java
#	src/main/java/ie/atu/userinterface/GPU/GPUController.java
#	src/main/java/ie/atu/userinterface/GPU/GPUHardwareClient.java
#	src/main/java/ie/atu/userinterface/GPU/GPUService.java
#	src/main/resources/templates/admin.html

+++ src/main/java/ie/atu/userinterface/Admin/AdminControllerNew.java
@@ -2,6 +2,7 @@ package ie.atu.userinterface.Admin;
 
 import ie.atu.userinterface.CPU.CPU;
 import ie.atu.userinterface.CPU.CPUService;
+import ie.atu.userinterface.GPU.GPU;
 import ie.atu.userinterface.GPU.GPUService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,36 +28,68 @@ public class AdminControllerNew {
     @GetMapping
     public String admin(Model model, HttpSession httpSession) {
         if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
-        model.addAttribute("componentSelected", "CPU");
         model.addAttribute("CPUs", cpuService.getCPU(null, null, null));
-        model.addAttribute("GPUs", gpuService.getGPU(null, null, null));
         return "admin";
     }
 
-//    @GetMapping("/cpus")
-//    public String getCPUs(@RequestParam String name, @RequestParam String brand, @RequestParam Float price) {
-//        List<CPU> CPUs = cpuService.getCPU(name, brand, price);
-//        return "redirect:/admin";
-//    }
+    //CPU ADMIN FUNCTIONS
+
+    @GetMapping("/cpus")
+    public String getCPUs(@RequestParam String name, @RequestParam String brand, @RequestParam Float price, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
+        List<CPU> CPUs = cpuService.getCPU(name, brand, price);
+        return "redirect:/admin";
+    }
 
     @PostMapping("/cpus")
-    public String createCPU(@ModelAttribute CPU cpu) {
+    public String createCPU(@ModelAttribute CPU cpu, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
         cpuService.saveCPU(cpu);
         return "redirect:/admin";
     }
 
     @DeleteMapping("/cpus/{id}")
-    public String deleteCPU(@PathVariable("id") Long id) {
+    public String deleteCPU(@PathVariable("id") Long id, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
         cpuService.deleteCPU(id);
         return "redirect:/admin";
     }
 
-    @PostMapping("/component-selection")
-    public String componentSelection(HttpSession httpSession) {
-
-
+    @PutMapping("/cpus/{id}")
+    public String editCPU(@PathVariable("id") Long id, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
+        cpuService.editCPU(id);
         return "redirect:/admin";
+    }
 
+    //GPU ADMIN FUNCTIONS
+
+    @GetMapping("/gpus")
+    public String getGPUs(HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
+        List<GPU> GPUs = gpuService.getGPU();
+        return "redirect:/admin";
+    }
+
+    @PostMapping("/cpus")
+    public String createGPU(@ModelAttribute GPU gpu, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
+        gpuService.saveGPU(gpu);
+        return "redirect:/admin";
+    }
+
+    @DeleteMapping("/cpus/{id}")
+    public String deleteGPU(@PathVariable("id") Long id, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
+        gpuService.deleteGPU(id);
+        return "redirect:/admin";
+    }
+
+    @PutMapping("/cpus/{id}")
+    public String editGPU(@PathVariable("id") Long id, HttpSession httpSession) {
+        if (httpSession.getAttribute("loggedIn").equals(false)) return "redirect:/";
+        gpuService.editGPU(id);
+        return "redirect:/admin";
     }
 
 }
